Fix warnings on Visual Studio (#1456)
* Fix warning C4244 in src/layer/convolution.cpp

C4244: '=': conversion from 'double' to 'float', possible loss of data

* Fix warning C4244 in src/layer/convolution_sgemm_int8.h

C4244: 'initializing': conversion from 'double' to 'int', possible loss of data

* Fix warning C4244 in src/layer/deconvolution.cpp

C4244: '=': conversion from 'double' to 'float', possible loss of data

* Fix warning C4244 in src/layer/elu.cpp

C4244: '=': conversion from 'double' to 'float', possible loss of data

* Fix warning C4267 in src/layer/embed.cpp

C4267: 'initializing': conversion from 'size_t' to 'int', possible loss of data

* Fix warning C4244 in src/layer/exp.cpp

C4244: '=': conversion from 'double' to 'float', possible loss of data

* Fix warning C4244 in src/layer/innerproduct.cpp

C4244: '=': conversion from 'double' to 'float', possible loss of data

* Fix warning C4244 in src/layer/log.cpp

C4244: '=': conversion from 'double' to 'float', possible loss of data
C4244: 'initializing': conversion from 'double' to 'float', possible loss of data

* Fix warning C4244 in src/layer/lrn.cpp

C4244: '=': conversion from 'double' to 'float', possible loss of data

* Fix warning C4244 in src/layer/mvn.cp

C4244: 'initializing': conversion from 'double' to 'float', possible loss of data

* Fix warning C4244 in src/layer/power.cpp

C4244: '=': conversion from 'double' to 'float', possible loss of data

* Fix warnings C4244 and C4267 in src/layer/proposal.cpp

C4244: 'initializing': conversion from 'double' to 'float', possible loss of data
C4244: 'initializing': conversion from 'double' to 'int', possible loss of data
C4267: 'argument': conversion from 'size_t' to 'int', possible loss of data
C4267: 'initializing': conversion from 'size_t' to 'int', possible loss of data

* Fix warning C4244 in src/layer/reduction.cpp

C4244: 'return': conversion from 'double' to 'T', possible loss of data

* Fix warning C4244 in src/layer/tanh.cpp

C4244: '=': conversion from 'double' to 'float', possible loss of data

* Fix warning C4244 in src/layer/binaryop.cpp

C4244: '=': conversion from 'double' to 'float', possible loss of data

* Fix warnings C4244 and C4267 in src/layer/unaryop.cpp

C4244: 'return': conversion from 'double' to 'T', possible loss of data
C4267: 'initializing': conversion from 'size_t' to 'int', possible loss of data

* Fix warning C4244 in src/layer/x86/convolutiondepthwise_3x3_int8.h

C4244: 'initializing': conversion from 'double' to 'int', possible loss of data

+++ src/layer/embed.cpp
@@ -53,7 +53,7 @@ int Embed::load_model(const ModelBin& mb)
 
 int Embed::forward(const Mat& bottom_blob, Mat& top_blob, const Option& opt) const
 {
-    int words = bottom_blob.total();
+    int words = static_cast<int>(bottom_blob.total());
 
     top_blob.create(num_output, words, 4u, opt.blob_allocator);
     if (top_blob.empty())
